aarch64 and related

+++ hal/armv7a/imx6ull/timer.c
@@ -139,7 +139,7 @@ int hal_timerRegister(intrFn_t f, void *data, intr_handler_t *h)
 }
 
 
-char *hal_timerFeatures(char *features, unsigned int len)
+char *hal_timerFeatures(char *features, size_t len)
 {
 	(void)hal_strncpy(features, "Using EPIT and GPT timers", len);
 	features[len - 1U] = '\0';
