+++ hal/armv7r/tda4vm/interrupts.c
@@ -198,7 +198,7 @@ int hal_interruptsSetHandler(intr_handler_t *h)
 }
 
 
-char *hal_interruptsFeatures(char *features, unsigned int len)
+char *hal_interruptsFeatures(char *features, size_t len)
 {
 	hal_strncpy(features, "Using VIM interrupt controller", len);
 	features[len - 1] = 0;
