+++ hal/interrupts.h
@@ -25,7 +25,7 @@ int hal_interruptsSetHandler(intr_handler_t *h);
 int hal_interruptsDeleteHandler(intr_handler_t *h);
 
 
-char *hal_interruptsFeatures(char *features, unsigned int len);
+char *hal_interruptsFeatures(char *features, size_t len);
 
 
 void _hal_interruptsInit(void);
