+++ hal/sparcv8leon/gaisler/irqmp.c
@@ -186,7 +186,7 @@ int hal_interruptsDeleteHandler(intr_handler_t *h)
 }
 
 
-char *hal_interruptsFeatures(char *features, unsigned int len)
+char *hal_interruptsFeatures(char *features, size_t len)
 {
 	(void)hal_strncpy(features, "Using IRQMP interrupt controller", len);
 	features[len - 1U] = '\0';
